requirements.txt
# --- CORE: Document & Tokenization ---
python-docx==1.1.0

# --- LLM & TOKENIZATION ---
torch==2.2.2
transformers==4.45.2
accelerate
bitsandbytes

# --- Embeddings & Vector Store ---
sentence-transformers==2.3.1
faiss-cpu

# --- Data & Viz ---
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1

# --- Utils ---
python-dotenv==1.0.0
tqdm==4.66.1
jupyter==1.0.0
pytest==7.4.3

# Python Version: 3.12.x (empfohlen: 3.12.7)